+++ src/pick_keys.d.ts
@@ -1,5 +1,5 @@
 /**
  * Picks the keys on `in_key_obj` from `obj`
  */
-export declare function pick_keys(obj: object, in_key_obj: object): object;
+export declare function pick_keys<I extends object = object, KEYS_OBJ extends object = object>(obj: I, in_key_obj: KEYS_OBJ): Partial<I>;
 export { pick_keys as pick__keys };