+++ src/pick_keys.ts
@@ -3,8 +3,10 @@ import { keys } from './keys'
 /**
  * Picks the keys on `in_key_obj` from `obj`
  */
-export function pick_keys(obj: object, in_key_obj: object) {
-	return pick(obj, ...keys(in_key_obj))
+export function pick_keys<I extends object = object, KEYS_OBJ extends object = object>(
+	obj:I, in_key_obj:KEYS_OBJ
+) {
+	return pick(obj, ...keys(in_key_obj)) as Partial<I>
 }
 export {
 	pick_keys as pick__keys