+++ src/unpick.d.ts
@@ -1,5 +1,5 @@
-import type { maybe } from '@ctx-core/function';
 /**
  * Does not include `key_a1` from `obj`
  */
-export declare function unpick<I extends unknown = unknown>(obj: Record<string, I>, ...key_a1: string[]): maybe<Record<string, I>>;
+export declare function unpick<I extends object = object>(obj: I, ...key_a1: string[]): Partial<I>;
+export declare function maybe_unpick<I extends object = object>(obj: I, ...key_a1: string[]): void | Partial<I>;