+++ src/unpick_keys.d.ts
@@ -1,5 +1,5 @@
 /**
- * Does not include keys on `keys_o` from `rec`
+ * Does not include keys on `keys_obj` from `rec`
  */
-export declare function unpick_keys<I extends unknown = unknown>(rec: Record<string, I>, keys_o: Record<string, unknown>): Record<string, I>;
+export declare function unpick_keys<I extends object = object, KEYS_OBJ extends object = object>(rec: I, keys_obj: KEYS_OBJ): Partial<I>;
 export { unpick_keys as unpick__keys };