+++ src/unpick_keys.ts
@@ -1,13 +1,12 @@
 import { unpick } from './unpick'
 import { keys } from './keys'
 /**
- * Does not include keys on `keys_o` from `rec`
+ * Does not include keys on `keys_obj` from `rec`
  */
-export function unpick_keys<I extends unknown = unknown>(
-	rec:Record<string, I>,
-	keys_o:Record<string, unknown>
+export function unpick_keys<I extends object = object, KEYS_OBJ extends object = object>(
+	rec:I, keys_obj:KEYS_OBJ
 ) {
-	return unpick<I>(rec, ...keys(keys_o)) as Record<string, I>
+	return unpick<I>(rec, ...keys(keys_obj)) as Partial<I>
 }
 export {
 	unpick_keys as unpick__keys